Move QuantME utilities from plugin to core
